+++ module/cvm/doc/changelog.md
@@ -22,6 +22,8 @@ repository.
 - Change
     - Remove anything related to `CVMByte` (since removed from newer Convex version)
     - `$.std/softness` -> `$.std/ref-stat` (provides more insights that counting soft refs)
+    - `$.db/write` and `$.db/root-write` return a new version of the input cell
+        - See docstrings, the old version should be discarded in favor of that new version 
 
 
 ---